requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.2
tenacity==8.5.0
PyYAML==6.0.2
pydantic==2.9.1
playwright==1.47.0
playwright-stealth==1.0.6
rich==13.8.1
python-dotenv==1.0.1

